opa: split into 2 test files
